--- src/Main.java
@@ -1,7 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-        StringReverse sr = new StringReverse();
-        String a = sr.reverseString("Hello");
-        System.out.println(a);
-    }
-}
