+++ src/com/company/Main.java
@@ -1,0 +1,10 @@
+package com.company;
+
+public class Main {
+
+    public static void main(String[] args) {
+            StringReverse sr = new StringReverse();
+            String a = sr.reverseString("Hello");
+            System.out.println(a);
+    }
+}
